ControlSate와 ControlContextService 분리
그 외에 StructureFactory 사용하지 않고 오류를 일으키는 함수도 제거함
