+++ CPP/NativeLib/NativeLib/StructureFactory.h
@@ -8,9 +8,4 @@ protected:
 
 
 public:
-	WorldObject createObject(Transform2D transform, Vector2 scale) {
-		Structure new_tile = Structure("noname", transform, scale, A);
-
-		return new_tile;
-	}
 };
